requirements.txt
djangorestframework
dj-database-url
django-heroku
gunicorn
psycopg2
python-decouple
whitenoise
drf-writable-nested
Pillow
